Add Meals FireStore Saving Updating Deleteng

+++ Screens/Food/SubScene/Meal/Interactors/MealInteractor.swift
@@ -15,18 +15,27 @@ protocol MealBusinessLogic {
 
 class MealInteractor: MealBusinessLogic {
   
-  var presenter    : MealPresentationLogic?
-  var realmManager : MealRealmManager!
+  var presenter     : MealPresentationLogic?
+  var realmManager  : MealRealmManager!
+  var addService    : AddService!
+  var updateService : UpdateService!
+  var deleteService : DeleteService!
   
 //  var items: Results<MealRealm>!
   
   var isDefaultList = true
   
-  func makeRequest(request: Meal.Model.Request.RequestType) {
-    if realmManager == nil {
-      realmManager = MealRealmManager()
-    }
+  init() {
+    let locator = ServiceLocator.shared
     
+    realmManager  = locator.getService()
+    addService    = locator.getService()
+    updateService = locator.getService()
+    deleteService = locator.getService()
+  }
+  
+  func makeRequest(request: Meal.Model.Request.RequestType) {
+
     
     workWithMealRequest(request: request)
     workWithProductListFromMealRequests(request: request)
@@ -77,14 +86,29 @@ class MealInteractor: MealBusinessLogic {
     switch  request {
       // Product List Delete Update Add
       case .deleteProductFromMeal(let mealId, let productName):
+        
+        print("Delete Product From Realm and Firebase",productName)
+        
+        guard let productIdRemove = realmManager.getproductIdByNameFromMeal(mealId: mealId, productName: productName) else {return}
+        
         realmManager.deleteProductFromMeal(productName: productName, mealId: mealId) // realmDBDidChange
+        
+        
+        deleteProductFromMealFireStore(productId: productIdRemove, mealId: mealId)
       
-      case .updateProductPortionFromMeal(let mealId, let rowProduct, let portion):
-        realmManager.updateProductPortionFromMeal(row: rowProduct, mealId: mealId, portion: portion) // realmDBDidChange
+      case .updateProductPortionFromMeal(let mealId, let productName, let portion):
+        
+        realmManager.updateProductPortionFromMeal(productName: productName, mealId: mealId, portion: portion) // realmDBDidChange
+       
+        guard let productId = realmManager.getproductIdByNameFromMeal(mealId: mealId, productName: productName) else {return}
+        
+        updateProductsPortionInMealsFireStore(productId: productId, portion: portion, mealId: mealId)
       
       case .addProductInMeal(let mealId, let product):
-        realmManager.addproductInMeal(mealId: mealId, product: product) // realmDBDidChange
-      
+        
+        guard let newProduct = realmManager.addproductInMeal(mealId: mealId, product: product) else {return} // realmDBDidChange
+        addProductsInMealsFireStore(product: newProduct, mealId: mealId)
+        
       
     default:break
     }
@@ -104,24 +128,36 @@ class MealInteractor: MealBusinessLogic {
       
     // Expand Meal
       case .expandedMeal(let mealId):
+      
         realmManager.changeMealExpande(mealId: mealId) // realmDBDidChange
       // Delete Meal
       case .deleteMeal(let mealId):
         
         realmManager.deleteMeal(mealId: mealId) // realmDBDidChange
+       deleteMealFromFireStore(mealId: mealId)
       // Add or Update Meal
+      
       case .addOrUpdateNewMeal(let viewModel):
         
         if viewModel.mealId != nil {
-          // Обновляем продукт
+          
+          
           realmManager.updateAllMealFields(viewModel: viewModel) { [weak self] (success) in
             self?.presenter?.presentData(response: .passSuccessAddOrUpdateMeal(isSuccessAdd:success, isUpdateMeal: true)) // realmDBDidChange
           }
+          updateMealInFireStore(viewModel: viewModel)
+          
           
         } else {
-          realmManager.addNewMeal(viewModel: viewModel) { [weak self] (success) in
+          
+          let mealRealm = createMealRealm(viewModel: viewModel)
+          
+          realmManager.addNewMeal(mealRealm: mealRealm) { [weak self] (success) in
+            
             self?.presenter?.presentData(response: .passSuccessAddOrUpdateMeal(isSuccessAdd:success, isUpdateMeal: false)) // realmDBDidChange
           }
+          
+          addMealToFireStore(mealRealm: mealRealm)
           
         }
       
@@ -129,25 +165,142 @@ class MealInteractor: MealBusinessLogic {
     }
   }
   
-  // for NewMealViewModel Presenter
-  
-  private func getUpdateMeal(mealId: String?) -> MealRealm? {
-    
-    var updateMeal: MealRealm?
-    if let mealId = mealId {
-      updateMeal = realmManager.getMealById(mealId: mealId)
-    }
-    return updateMeal
-  }
-  
-  private func getListOfTypeOfMeal() -> [String] {
-    return realmManager.getTypeOFMealList()
-  }
+ 
   
   private func realmDBDidChange(items: Results<MealRealm>) {
-    print("Change DB delete productsIn Realm")
+    print("Change DB Meal Realm")
     
     presenter?.presentData(response: .prepareMealViewModelforTableView(items: items, bySections: isDefaultList))
   }
   
+}
+
+
+// MARK: Work with FireStore
+
+extension MealInteractor {
+  
+  // MARK: Products from Meals
+  
+  // Delete Product
+  
+  private func deleteProductFromMealFireStore(productId: String, mealId: String) {
+    
+    deleteService.deleteProductFromMealFireStore(mealId: mealId, productId: productId)
+  }
+  
+  // Add Product
+  private func addProductsInMealsFireStore(product:ProductRealm,mealId: String) {
+    
+    let productFireStore = createNewFireBaseProduct(productRealm: product)
+    
+    addService.addProductInMeal(mealId: mealId, product: productFireStore)
+    
+  }
+  // Update Product
+  private func updateProductsPortionInMealsFireStore(
+    productId : String,
+    portion   : Int,
+    mealId    : String) {
+    
+    
+    
+    updateService.updateProductInMealInFireStore(productId: productId, mealId: mealId, portion: portion)
+  }
+  
+  
+  
+  // MARK: Meals
+  private func deleteMealFromFireStore(mealId: String) {
+    deleteService.deleteMealFromFireStore(mealId: mealId)
+  }
+  
+  
+  // UPDATE Meal
+  private func updateMealInFireStore(viewModel: MealViewModel) {
+    
+    let data = createMealData(viewModel: viewModel)
+    
+    updateService.updateMealInFireStore(
+      mealId: viewModel.mealId!,
+      data: data)
+  }
+  
+  private func createMealData(viewModel: MealViewModel) -> [String: Any] {
+    
+    [
+      MealNetworkModel.CodingKeys.name.rawValue         : viewModel.name,
+      MealNetworkModel.CodingKeys.isExpandMeal.rawValue : viewModel.isExpanded,
+      MealNetworkModel.CodingKeys.typeMeal.rawValue     : viewModel.typeMeal,
+      MealNetworkModel.CodingKeys.listProduct.rawValue  : viewModel.products
+    ]
+    
+  }
+  
+   private func createNewFireBaseProduct(productRealm: ProductRealm) -> ProductNetworkModel {
+    
+    print(productRealm,"Product Realm")
+
+     return ProductNetworkModel(
+       id                     : productRealm.id,
+       name                   : productRealm.name,
+       category               : productRealm.category,
+       carboIn100grm          : productRealm.carboIn100grm,
+       portion                : productRealm.portion,
+       percantageCarboInMeal  : productRealm.percantageCarboInMeal,
+       userSetInsulinOnCarbo  : productRealm.userSetInsulinOnCarbo,
+       insulinOnCarboToML     : productRealm.insulinOnCarboToML,
+       isFavorits             : productRealm.isFavorits)
+   }
+  
+  
+  
+  // ADD Meal
+  private func addMealToFireStore(mealRealm:MealRealm) {
+    
+    let mealNetworkModel = createMealNetworkModel(mealRealm: mealRealm)
+    
+    addService.addMealToFireStore(meal: mealNetworkModel)
+  }
+  
+  private func createMealNetworkModel(mealRealm: MealRealm) -> MealNetworkModel {
+    
+    MealNetworkModel(
+      id           : mealRealm.id,
+      isExpandMeal : mealRealm.isExpandMeal,
+      name         : mealRealm.name,
+      typeMeal     : mealRealm.typeMeal,
+      listProduct  : [])
+  }
+  
+}
+
+// MARK: Work with Realm Base
+
+extension MealInteractor {
+  
+  
+  private func createMealRealm(viewModel: MealViewModel) -> MealRealm {
+    let newMeal = MealRealm(
+    
+    name         : viewModel.name,
+    typeMeal     : viewModel.typeMeal,
+    isExpandMeal : viewModel.isExpanded)
+    return newMeal
+  }
+  
+  // MARK: Get Update Meal
+   
+   private func getUpdateMeal(mealId: String?) -> MealRealm? {
+     
+     var updateMeal: MealRealm?
+     if let mealId = mealId {
+       updateMeal = realmManager.getMealById(mealId: mealId)
+     }
+     return updateMeal
+   }
+   
+   private func getListOfTypeOfMeal() -> [String] {
+     return realmManager.getTypeOFMealList()
+   }
 }
